// src/app/models/ordens.ts
export interface Ordens {
    id?:                  any;
    dataAbertura?:      String;
    dataFechamento?:    String;
    prioridade:         String;
    status:             String;
    titulo:             String;
    descricao:          String;
    tecnico:               any;
    cliente:               any;
    nomeCliente:        String;
    nomeTecnico:        String;
}